Force high quality only when speaking
The high wuality video was always forced if the local audio is enabled.
However, if the local participant has the audio enabled but is not
speaking she will not be the promoted participant (except if she was the
last one that spoke, but that should be just a temporary situation).
Therefore now the high quality is forced only if the local participant
is speaking or has just finished speaking.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>