Increase constraints max value if applying the default ones failed
Some browsers may fail to apply a constraint when a max attribute is
set (even if once the max attribute is increased the constraint is set
to a much lower value). Thefore now, when applying the default
constraints fail, the constraints are tried to be applied again with
increments in their max value until it succeeds (or finally fails).

Besides that the increment is done separately for resolution and frame
rate. First the resolution is set and, once succeeded, the frame rate is
then set, as it seems that browsers handle better setting the resolution
than the frame rate.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>